+++ relay/kusama/src/lib.rs
@@ -51,12 +51,11 @@ use frame_support::{
 	parameter_types,
 	traits::{
 		fungible::HoldConsideration,
-		schedule::DispatchTime,
 		tokens::{imbalance::ResolveTo, UnityOrOuterConversion},
-		ConstU32, ConstU8, ConstUint, Currency, DefensiveResult, EitherOf, EitherOfDiverse,
-		EnsureOrigin, EnsureOriginWithArg, Equals, FromContains, InstanceFilter,
-		KeyOwnerProofSystem, LinearStoragePrice, OnUnbalanced, PrivilegeCmp, ProcessMessage,
-		ProcessMessageError, WithdrawReasons,
+		ConstU32, ConstU8, ConstUint, Currency, EitherOf, EitherOfDiverse, EnsureOrigin,
+		EnsureOriginWithArg, Equals, FromContains, InstanceFilter, KeyOwnerProofSystem,
+		LinearStoragePrice, OnUnbalanced, PrivilegeCmp, ProcessMessage, ProcessMessageError,
+		WithdrawReasons,
 	},
 	weights::{
 		constants::{WEIGHT_PROOF_SIZE_PER_KB, WEIGHT_REF_TIME_PER_MICROS},
@@ -128,8 +127,8 @@ use sp_runtime::{
 		Get, IdentityLookup, Keccak256, OpaqueKeys, SaturatedConversion, Saturating, Verify,
 	},
 	transaction_validity::{TransactionPriority, TransactionSource, TransactionValidity},
-	AccountId32, ApplyExtrinsicResult, FixedU128, KeyTypeId, OpaqueValue, Perbill, Percent,
-	Permill, RuntimeDebug,
+	ApplyExtrinsicResult, FixedU128, KeyTypeId, OpaqueValue, Perbill, Percent, Permill,
+	RuntimeDebug,
 };
 use sp_staking::{EraIndex, SessionIndex};
 #[cfg(any(feature = "std", test))]
@@ -1913,57 +1912,6 @@ const AH_MAXIMUM_BLOCK_WEIGHT: Weight = Weight::from_parts(
 	polkadot_primitives::MAX_POV_SIZE as u64,
 );
 
-/// Kusama multisig members for the `Manager` privilege of the RC Migrator pallet with threshold 3.
-///
-/// ACCOUNTS MUST BE ABLE TO SIGN VIA POLKADOTJS APPS `DEVELOPER->SIGN AND VERIFY` FEATURE.
-/// This does *not* work for signing devices and implies that the account is *hot*. The account
-/// does not need to have a balance and the chain does not matter.
-///
-/// Will be used to respond to issues during the Asset Hub Migration and to adjust the scheduled
-/// timepoint to ensure that it runs at the right time. Most members do not need to do anything
-/// but are just in place to act as emergency backup contacts.
-#[cfg(feature = "std")]
-fn multisig_members() -> Vec<AccountId32> {
-	use sp_core::crypto::Ss58Codec;
-
-	let addresses = vec![
-		"5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY", // Alice SR25519
-		"5FA9nQDVg267DEd8m1ZypXLBnvN7SFxYwV7ndqSYGiN9TTpu", // Alice ED25519
-		"5C7C2Z5sWbytvHpuLTvzKunnnRwQxft1jiqrLD5rhucQ5S9X", /* Alice ECDSA Address (not SS58
-		                                                     * Public Key) */
-		"FoQJpPyadYccjavVdTWxpxU7rUEaYhfLCPwXgkfD6Zat9QP", // Bob
-		"Fr4NzY1udSFFLzb2R3qxVQkwz9cZraWkyfH4h3mVVk7BK7P", // Charlie
-		"HnMAUz7r2G8G3hB27SYNyit5aJmh2a5P4eMdDtACtMFDbam", // Eve
-	];
-
-	addresses
-		.into_iter()
-		.filter_map(|ss| sp_runtime::AccountId32::from_ss58check(ss).defensive_ok())
-		.collect()
-}
-
-#[cfg(not(feature = "std"))]
-fn multisig_members() -> Vec<AccountId32> {
-	use sp_core::crypto::Ss58Codec;
-
-	let addresses = vec![
-		"FFFF3gBSSDFSvK2HBq4qgLH75DHqXWPHeCnR1BSksAMacBs", /* Basti Kusama Fellowship https://kusama.subscan.io/account/FFFF3gBSSDFSvK2HBq4qgLH75DHqXWPHeCnR1BSksAMacBs */
-		"FcxNWVy5RESDsErjwyZmPCW6Z8Y3fbfLzmou34YZTrbcraL", /* Gav Kusama Fellowship https://kusama.subscan.io/account/FcxNWVy5RESDsErjwyZmPCW6Z8Y3fbfLzmou34YZTrbcraL */
-		"HL8bEp8YicBdrUmJocCAWVLKUaR2dd1y6jnD934pbre3un1", /* Kian Kusama Fellowship https://kusama.subscan.io/account/HL8bEp8YicBdrUmJocCAWVLKUaR2dd1y6jnD934pbre3un1 */
-		"F2jgWXy7X8GQ2ykf1UGrsCXRERZvjEcd2aDf39fMf3BWVy6", /* Oliver Hot https://kusama.subscan.io/account/F2jgWXy7X8GQ2ykf1UGrsCXRERZvjEcd2aDf39fMf3BWVy6 */
-		"GcDZZCVPwkPqoWxx8vfLb4Yfpz9yQ1f4XEyqngSH8ygsL9p", /* Joe Kusama Fellowship https://kusama.subscan.io/account/GcDZZCVPwkPqoWxx8vfLb4Yfpz9yQ1f4XEyqngSH8ygsL9p */
-		"12HWjfYxi7xt7EvpTxUis7JoNWF7YCqa19JXmuiwizfwJZY2", /* Muharem Polkadot Fellowship https://polkadot.subscan.io/account/12HWjfYxi7xt7EvpTxUis7JoNWF7YCqa19JXmuiwizfwJZY2 */
-		"121dd6J26VUnBZ8BqLGjANWkEAXSb9mWq1SB7LsS9QNTGFvz", /* Adrian Polkadot Fellowship https://polkadot.subscan.io/account/121dd6J26VUnBZ8BqLGjANWkEAXSb9mWq1SB7LsS9QNTGFvz */
-		"12pRzYaysQz6Tr1e78sRmu9FGB8gu8yTek9x6xwVFFAwXTM8", /* RobK Polkadot Fellowship https://polkadot.subscan.io/account/12pRzYaysQz6Tr1e78sRmu9FGB8gu8yTek9x6xwVFFAwXTM8 */
-		"FcJnhk4i1bfuN9E2B6yMnL8h97ogtuL7e4ZpqnYgvj9moQy", /* Donal Polkadot Fellowship https://kusama.subscan.io/account/FcJnhk4i1bfuN9E2B6yMnL8h97ogtuL7e4ZpqnYgvj9moQy */
-	];
-
-	addresses
-		.into_iter()
-		.filter_map(|ss| sp_runtime::AccountId32::from_ss58check(ss).defensive_ok())
-		.collect()
-}
-
 parameter_types! {
 	// Exvivalent to `kusama_asset_hub_runtime::MessageQueueServiceWeight`.
 	pub AhMqServiceWeight: Weight = Perbill::from_percent(50) * AH_MAXIMUM_BLOCK_WEIGHT;
@@ -1973,9 +1921,6 @@ parameter_types! {
 	pub AhExistentialDeposit: Balance = EXISTENTIAL_DEPOSIT / 100;
 	pub const XcmResponseTimeout: BlockNumber = 30 * DAYS;
 	pub const AhUmpQueuePriorityPattern: (BlockNumber, BlockNumber) = (18, 2);
-
-	pub MultisigMembers: Vec<AccountId32> = multisig_members();
-
 }
 
 pub struct ProxyTypeAny;
@@ -2021,6 +1966,7 @@ impl pallet_rc_migrator::Config for Runtime {
 	type RecoveryBlockNumberProvider = System;
 	type MultisigMembers = ();
 	type MultisigThreshold = ConstU32<{ u32::MAX }>;
+	type MultisigMaxVotesPerRound = (); // defunct
 }
 
 construct_runtime! {
@@ -2199,51 +2145,15 @@ pub type TxExtension = (
 #[allow(deprecated, missing_docs)]
 pub mod migrations {
 	use super::*;
-	use frame_support::traits::OnRuntimeUpgrade;
-	use pallet_rc_migrator::{MigrationStage, MigrationStartBlock, RcMigrationStage};
 
 	/// Unreleased migrations. Add new ones here:
-	pub type Unreleased = (KickOffAhm<Runtime>,);
+	pub type Unreleased = ();
 
 	/// Migrations/checks that do not need to be versioned and can run on every update.
 	pub type Permanent = pallet_xcm::migration::MigrateToLatestXcmVersion<Runtime>;
 
 	/// All migrations that will run on the next runtime upgrade.
 	pub type SingleBlockMigrations = (Unreleased, Permanent);
-
-	/// Kick off the Asset Hub Migration.
-	pub struct KickOffAhm<T>(pub core::marker::PhantomData<T>);
-	impl<T: pallet_rc_migrator::Config> OnRuntimeUpgrade for KickOffAhm<T> {
-		fn on_runtime_upgrade() -> Weight {
-			if MigrationStartBlock::<T>::exists() ||
-				RcMigrationStage::<T>::get() != MigrationStage::Pending
-			{
-				// Already started or scheduled
-				log::info!("KickOffAhm: Asset Hub Migration already started or scheduled");
-				return T::DbWeight::get().reads(2)
-			}
-
-			let result = pallet_rc_migrator::Pallet::<T>::do_schedule_migration(
-				// Migration start block, Tuesday 7th Oct 8 AM UTC
-				// Naive estimate [30422513](https://kusama.subscan.io/block/30422513)
-				// Adjusted with -20.52 clock skew / day [30423691](https://kusama.subscan.io/block/30423691), avg delay per day over the last 30 days.
-				DispatchTime::At(30423691u32.into()),
-				// Warm up to wait for Messaging queues to empty
-				DispatchTime::After((5 * MINUTES).into()),
-				// Cool off to verify the success of the migration
-				DispatchTime::After((60 * MINUTES).into()),
-				// Respect the session scheduling check:
-				Default::default(),
-			);
-			if let Err(e) = result {
-				log::error!("KickOffAhm: Failed to schedule Asset Hub Migration: {e:?}");
-			} else {
-				log::info!("KickOffAhm: Scheduled Asset Hub Migration");
-			}
-
-			T::DbWeight::get().reads_writes(1, 1)
-		}
-	}
 }
 
 /// Unchecked extrinsic type as expected by this runtime.
@@ -3249,273 +3159,6 @@ sp_api::impl_runtime_apis! {
 
 			Ok(batches)
 		}
-	}
-}
-
-#[cfg(test)]
-mod ahm_multisig {
-	use super::*;
-	use pallet_rc_migrator::ManagerMultisigVote;
-	use sp_core::Pair;
-	use sp_io::TestExternalities;
-	use sp_runtime::{
-		traits::{Dispatchable, ValidateUnsigned},
-		MultiSignature, MultiSigner,
-	};
-
-	#[test]
-	fn all_ss58s_decode() {
-		// ensure all non-dev account ids we have are valid ss58s
-		assert_eq!(MultisigMembers::get().len(), 6);
-	}
-
-	#[test]
-	#[ignore]
-	fn unsigned_manager_multisig_works() {
-		TestExternalities::default().execute_with(|| {
-			let call = pallet_rc_migrator::Call::<Runtime>::force_set_stage {
-				stage: Box::new(pallet_rc_migrator::MigrationStage::Starting),
-			};
-			let runtime_call = RuntimeCall::from(call.clone());
-			let other_call = pallet_rc_migrator::Call::<Runtime>::force_set_stage {
-				stage: Box::new(pallet_rc_migrator::MigrationStage::SchedulerMigrationInit),
-			};
-			let other_runtime_call = RuntimeCall::from(other_call.clone());
-
-			// initial round is zero
-			assert_eq!(pallet_rc_migrator::ManagerMultisigRound::<Runtime>::get(), 0);
-
-			{
-				// Ferdie is not part of the multisig, will get rejected on validate
-				let ferdie = sp_keyring::Sr25519Keyring::Ferdie.pair();
-				let payload = ManagerMultisigVote::new(
-					MultiSigner::Sr25519(ferdie.public()),
-					runtime_call.clone(),
-					0,
-				);
-				let sig = ferdie.sign(payload.encode_with_bytes_wrapper().as_ref());
-				let call = pallet_rc_migrator::Call::<Runtime>::vote_manager_multisig {
-					payload: Box::new(payload),
-					sig: MultiSignature::Sr25519(sig),
-				};
-				assert!(pallet_rc_migrator::Pallet::<Runtime>::validate_unsigned(
-					TransactionSource::External,
-					&call
-				)
-				.is_err());
-			}
-
-			{
-				// Alice signs a wrong message, rejected
-				let alice = sp_keyring::Sr25519Keyring::Alice.pair();
-				let payload = ManagerMultisigVote::new(
-					MultiSigner::Sr25519(sp_keyring::Sr25519Keyring::Bob.pair().public()),
-					runtime_call.clone(),
-					0,
-				);
-				let sig = alice.sign(payload.encode_with_bytes_wrapper().as_ref());
-				let call = pallet_rc_migrator::Call::<Runtime>::vote_manager_multisig {
-					payload: Box::new(payload),
-					sig: MultiSignature::Sr25519(sig),
-				};
-				assert!(
-					pallet_rc_migrator::Pallet::<Runtime>::validate_unsigned(
-						TransactionSource::External,
-						&call
-					)
-					.is_err(),
-					"Alice signs a wrong message, rejected"
-				);
-			}
-
-			let alice_sig_for_first_round = {
-				// Alice signs, not executed yet
-				let alice = sp_keyring::Sr25519Keyring::Alice.pair();
-				let payload = ManagerMultisigVote::new(
-					MultiSigner::Sr25519(alice.public()),
-					runtime_call.clone(),
-					0,
-				);
-				let sig = alice.sign(payload.encode_with_bytes_wrapper().as_ref());
-				let call = pallet_rc_migrator::Call::<Runtime>::vote_manager_multisig {
-					payload: Box::new(payload),
-					sig: MultiSignature::Sr25519(sig),
-				};
-
-				assert!(pallet_rc_migrator::Pallet::<Runtime>::validate_unsigned(
-					TransactionSource::External,
-					&call
-				)
-				.is_ok());
-				assert!(RuntimeCall::from(call).dispatch(RuntimeOrigin::none()).is_ok());
-				assert_eq!(
-					pallet_rc_migrator::ManagerMultisigs::<Runtime>::get(runtime_call.clone())
-						.len(),
-					1
-				);
-
-				MultiSignature::Sr25519(sig)
-			};
-
-			frame_support::hypothetically!({
-				// Alice ED25519 signs
-				let alice = sp_keyring::Ed25519Keyring::Alice.pair();
-				let payload = ManagerMultisigVote::new(
-					MultiSigner::Ed25519(sp_keyring::Ed25519Keyring::Alice.pair().public()),
-					runtime_call.clone(),
-					0,
-				);
-				let sig = alice.sign(payload.encode_with_bytes_wrapper().as_ref());
-				let call = pallet_rc_migrator::Call::<Runtime>::vote_manager_multisig {
-					payload: Box::new(payload),
-					sig: MultiSignature::Ed25519(sig),
-				};
-				assert!(
-					pallet_rc_migrator::Pallet::<Runtime>::validate_unsigned(
-						TransactionSource::External,
-						&call
-					)
-					.is_ok(),
-					"Alice ED25519 signs"
-				);
-			});
-
-			frame_support::hypothetically!({
-				// Alice ECDSA Address signs
-				let seed = hex_literal::hex!(
-					"cb6df9de1efca7a3998a8ead4e02159d5fa99c3e0d4fd6432667390bb4726854"
-				);
-				let alice = sp_application_crypto::ecdsa::Pair::from_seed_slice(&seed).unwrap();
-				let payload = ManagerMultisigVote::new(
-					MultiSigner::Ecdsa(alice.public()),
-					runtime_call.clone(),
-					0,
-				);
-
-				let sig = alice.sign(payload.encode_with_bytes_wrapper().as_ref());
-				let call = pallet_rc_migrator::Call::<Runtime>::vote_manager_multisig {
-					payload: Box::new(payload),
-					sig: MultiSignature::Ecdsa(sig),
-				};
-				assert!(
-					pallet_rc_migrator::Pallet::<Runtime>::validate_unsigned(
-						TransactionSource::External,
-						&call
-					)
-					.is_ok(),
-					"Alice ECDSA Address signs"
-				);
-			});
-
-			{
-				// Bob signs, still waiting
-				let bob = sp_keyring::Sr25519Keyring::Bob.pair();
-				let payload = ManagerMultisigVote::new(
-					MultiSigner::Sr25519(bob.public()),
-					runtime_call.clone(),
-					0,
-				);
-				let sig = bob.sign(payload.encode_with_bytes_wrapper().as_ref());
-				let call = pallet_rc_migrator::Call::<Runtime>::vote_manager_multisig {
-					payload: Box::new(payload),
-					sig: MultiSignature::Sr25519(sig),
-				};
-				assert!(pallet_rc_migrator::Pallet::<Runtime>::validate_unsigned(
-					TransactionSource::External,
-					&call
-				)
-				.is_ok());
-				assert!(RuntimeCall::from(call).dispatch(RuntimeOrigin::none()).is_ok());
-				assert_eq!(
-					pallet_rc_migrator::ManagerMultisigs::<Runtime>::get(runtime_call.clone())
-						.len(),
-					2
-				);
-			}
-
-			{
-				// charlie signs something else, stored but not used.
-				let charlie = sp_keyring::Sr25519Keyring::Charlie.pair();
-				let payload = ManagerMultisigVote::new(
-					MultiSigner::Sr25519(charlie.public()),
-					other_runtime_call.clone(),
-					0,
-				);
-				let sig = charlie.sign(payload.encode_with_bytes_wrapper().as_ref());
-				let call = pallet_rc_migrator::Call::<Runtime>::vote_manager_multisig {
-					payload: Box::new(payload),
-					sig: MultiSignature::Sr25519(sig),
-				};
-				assert!(pallet_rc_migrator::Pallet::<Runtime>::validate_unsigned(
-					TransactionSource::External,
-					&call
-				)
-				.is_ok());
-				assert!(RuntimeCall::from(call).dispatch(RuntimeOrigin::none()).is_ok());
-				// 1 vote for the new one
-				assert_eq!(
-					pallet_rc_migrator::ManagerMultisigs::<Runtime>::get(
-						other_runtime_call.clone()
-					)
-					.len(),
-					1
-				);
-				// still 2 votes for this
-				assert_eq!(
-					pallet_rc_migrator::ManagerMultisigs::<Runtime>::get(runtime_call.clone())
-						.len(),
-					2
-				);
-			}
-
-			{
-				// eve signs, dispatched
-				let eve = sp_keyring::Sr25519Keyring::Eve.pair();
-				let payload = ManagerMultisigVote::new(
-					MultiSigner::Sr25519(eve.public()),
-					runtime_call.clone(),
-					0,
-				);
-				let sig = eve.sign(payload.encode_with_bytes_wrapper().as_ref());
-				let call = pallet_rc_migrator::Call::<Runtime>::vote_manager_multisig {
-					payload: Box::new(payload),
-					sig: MultiSignature::Sr25519(sig),
-				};
-				assert!(pallet_rc_migrator::Pallet::<Runtime>::validate_unsigned(
-					TransactionSource::External,
-					&call
-				)
-				.is_ok());
-				assert!(RuntimeCall::from(call).dispatch(RuntimeOrigin::none()).is_ok());
-				// executed, so no more votes
-				assert_eq!(
-					pallet_rc_migrator::RcMigrationStage::<Runtime>::get(),
-					pallet_rc_migrator::MigrationStage::Starting
-				);
-				// votes erased
-				assert!(pallet_rc_migrator::ManagerMultisigs::<Runtime>::get(runtime_call.clone())
-					.is_empty(),);
-				// round incremented
-				assert_eq!(pallet_rc_migrator::ManagerMultisigRound::<Runtime>::get(), 1);
-			}
-
-			{
-				// Alice's signature from first round cannot be re-used
-				let call = pallet_rc_migrator::Call::<Runtime>::vote_manager_multisig {
-					payload: Box::new(ManagerMultisigVote::new(
-						MultiSigner::Sr25519(sp_keyring::Sr25519Keyring::Alice.pair().public()),
-						runtime_call.clone(),
-						0,
-					)),
-					sig: alice_sig_for_first_round,
-				};
-				assert!(pallet_rc_migrator::Pallet::<Runtime>::validate_unsigned(
-					TransactionSource::External,
-					&call
-				)
-				.is_err());
-			}
-		})
 	}
 }